+++ template-parts/sidebar.php
@@ -2,7 +2,7 @@
 
 echo '<div id="sidebar_column">';
 
-	lawyerist_platinum_sponsors_widget();
+	lwyrst_plat_sponsors_widget();
 
 echo '</div>' . "\n" . '<!-- end #sidebar_column -->';